requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
pytz==2022.1
routers==0.10.1
sqlparse==0.4.2
